// apps/api/src/modules/sales/dto/query-sale.dto.d.ts
export declare class QuerySaleDto {
    page?: number;
    limit?: number;
    branchId?: string;
    clientId?: string;
    paymentStatus?: string;
    saleType?: string;
    fromDate?: string;
    toDate?: string;
    search?: string;
    isVoided?: boolean;
}
//# sourceMappingURL=query-sale.dto.d.ts.map